// src/yandex-fleet/common/yandex-fleet-common.type.ts
export enum YandexDriverProfessionType {
  TaxiDriver = 'taxi/driver',
  CargoCourierOnFoot = 'cargo/courier/on-foot',
  CargoCourierOnCar = 'cargo/courier/on-car',
  CargoCourierOnTruck = 'cargo/courier/on-truck',
  CargoCourierOnMotorcycle = 'cargo/courier/on-motorcycle',
}

export enum EmploymentType {
  SelfEmployed = 'selfemployed',
  ParkEmployee = 'park_employee',
  IndividualEntrepreneur = 'individual_entrepreneur',
}

export enum CountryCode {
  RU = 'rus',
  BY = 'blr',
  KZ = 'kaz',
  KG = 'kgz',
  TJ = 'tjk',
  UA = 'ukr',
  GE = 'geo',
}

export enum DriverWorkStatus {
  Working = 'working',
  NotWorking = 'not_working',
  Fired = 'fired',
}

export enum TaxSystemRus {
  OSN = 'OSN',
  USN = 'USN',
  AUSN = 'AUSN',
  OSNP = 'OSNP',
  USNP = 'USNP',
}